requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
pandas==2.2.2
openpyxl==3.1.5
python-dateutil==2.9.0.post0
